@@ -1,57 +1,56 @@
+[project]
+name = "planetae-orm"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+authors = [
+    { name = "Edmilson Monteiro Rodrigues Neto", email = "edmilson.rodriguesn38@gmail.com" }
+]
+requires-python = ">=3.13"
+dependencies = [
+    "aiomysql==0.2.0",
+    "aioodbc==0.5.0",
+    "aiosqlite==0.20.0",
+    "asyncpg==0.30.0",
+    "mariadb==1.1.11",
+    "motor==3.6.0",
+]
+
+[project.scripts]
+lint = "tools.lint:lint_and_format"
+test = "tools.test:test"
+docs = "tools.docs:build"
+
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
-[project]
-name = "planetae_db"
-version = "0.1.0"
-description = "A python package to integrate the main relationals databases and MongoDB asynchronously in an unified syntax."
-authors = [
-    {name = "Edmilson Monteiro Rodrigues Neto", email = "edmilson.rodriguesn38@gmail.com"}
+[dependency-groups]
+dev = [
+    "mypy>=1.13.0",
+    "ruff>=0.8.2",
 ]
-license = {text = "MIT"}
-readme = "README.md"
-requires-python = ">=3.12"
-keywords = ["database", "db", "asynchronous", "mongodb", "mariadb", "mysql", "sqlite3", "postgressql", "mssql"]
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: MIT License",
-    "Topic :: Database"
+tests = [
+    "pytest-asyncio>=0.24.0",
+    "pytest>=8.3.2",
+    "pytest-cov>=5.0.0",
+    "tox>=4.23.2",
 ]
-dependencies = [
-    "aiomysql>=0.2.0",
-    "mariadb>=1.1.10",
-    "aiosqlite>=0.20.0",
-    "asyncpg>=0.29.0",
-    "aioodbc>=0.5.0",
-    "motor>=3.5.1",
-    "planetae_logger>=0.3.0",
-    "flake8-pyproject>=1.2.3"
+docs = [
+    "autodoc>=0.5.0",
+    "sphinx>=8.1.3",
 ]
 
-[project.urls]
-repository = "https://github.com/EdmilsonRodrigues/planetae_db"
+[tool.ruff]
+line-length = 79
 
-[tool.hatch.build.targets.sdist.include]
-patterns = ["src/**/*"]
+target-version = "py313"
 
-[tool.pytest.ini_options]
-asyncio_default_fixture_loop_scope = "function"
-markers = [
-    "mariadb: runs only the tests related to the mariadb database. (Use the one you have installed)",
-    "mysql: runs only the tests related to the mysql database. (Use the one you have installed)",
-    "postgresql: runs only the tests related to the postgressql database. (Use the one you have installed)",
-    "mssql: runs only the tests related to the mssql database. (Use the one you have installed)",
-    "sqlite3: runs only the tests related to the sqlite3 database. (Use the one you have installed)",
-    "mongodb: runs only the tests related to the mongodb database. (Use the one you have installed)"
-]
-
-[tool.black]
-line-length = 120
-
-[tool.flake8]
-max-line-length = 120
-exclude = ["__init__.py"]
+[tool.ruff.lint]
+select = ["E", "F", "I", "UP"]
 
 
+[tool.ruff.format]
+docstring-code-format = true
+docstring-code-line-length = 72
 
